backend/requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.5
cryptography==3.2.1
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.6.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
mysqlclient==2.1.0
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pytz==2020.4
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing_extensions==4.0.1
whitenoise==5.3.0
wincertstore==0.2
